Moved price suggestion feature into a seperate app module
Keeping this logic in a seperated app module will reduce clutter in the listings module and make it easier to iterrate on this price suggestions feature in the future. Ideally in the future improvments would include refactoring this module to use or own database records and maybe also factor in the web scraped data. We might also think about caching or optimizing here.