Quick temp fix for registration next redirecting
I wanted to maintain the next success redirecting even when the user needs to register a new account. Ideally after registering the user would be authenticated automatically and sent directly to the next page rather than having to login manually. That will require refactoring to use a FormView instead of create view then  using the authenticate on login methods, so I decided just to do this for now.