Clean up - remove unused models from the unfinished refactor branch
These large commented out code sections were introduced by merging in a portion of the model refactor branch a while back. The planned refactor code still exists on the refactor/av-data-models branch. This wasn't finished in time to be fully functional for the final presentation.  A shame because it would've allow for a lot of intresting additional functionality.